Add ShelfCompletionChecker and integrate with CatSortMode
