+++ Assets/Assets/Scripts/ShelfCompletionChecker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eedd131669d442778c8654b4fb5c1f19
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
